requirements.txt
pyjwt >= 1.7.1
requests >= 2.21.0
cryptography >= 3.0